feat: update timer settings UI and precision
- Changed timer settings from PopupWindow to Dialog for better user experience
- Improved timer precision by changing unit from minutes to seconds
- Renamed menu_settimer.xml to dialog_set_timer.xml for clarity
- Renamed SetTimerWindow.kt to SetTimerDialog.kt for consistency

+++ app/src/main/java/com/hnidesu/timer/MainActivity.kt
@@ -69,7 +69,7 @@ class MainActivity : AppCompatActivity(), ServiceConnection {
         }
         mTimer.scheduleAtFixedRate(task, 0L, 500L)
         val viewHolder = ViewHolder(object : TaskOperationListener {
-            override fun onAddTask(packageName: String, timeout: Int): Boolean {
+            override fun onAddTask(packageName: String, timeout: Long): Boolean {
                 timerService.addTask(packageName, timeout)
                 return true
             }
